requirements.txt
# Telegram Bot
aiogram==3.3.0
python-dotenv==1.0.0

# Web Framework
Flask==3.0.0
Flask-CORS==4.0.0

# Database
psycopg2-binary==2.9.9

# Other
aiohttp==3.9.1
certifi==2023.11.17
